+++ src/hooks/useTransitionState.ts
@@ -1,4 +1,4 @@
-import { useCallback, useEffect, useRef, useState } from 'react'
+import { useCallback, useEffect, useId, useRef, useState } from 'react'
 
 export type TransitionState = 'opened' | 'closed' | 'opening' | 'closing'