+++ src/hooks/useValidators.ts
@@ -89,8 +89,8 @@ export const useTranslatedValidators = () => {
         return translation(result as 'outOfRangeString' | 'tooShort' | 'tooLong', { min, max })
       }
     },
-    email: (value: string) => {
-      const result = emailValidator(value)
+    email: (value: string | undefined) => {
+      const result = emailValidator(value ?? '')
       if (result) {
         return translation(result as 'invalidEmail')
       }
@@ -101,7 +101,7 @@ export const useTranslatedValidators = () => {
       if (result) {
         return translation(result as
             'outOfRangeSelectionItems' | 'tooFewSelectionItems' | 'tooManySelectionItems',
-          { min, max })
+        { min, max })
       }
     }
   }